+++ app/code/community/Nosto/Tagging/Model/Meta/Cart.php
@@ -59,13 +59,258 @@ class Nosto_Tagging_Model_Meta_Cart extends Mage_Core_Model_Abstract
         if (is_null($store)) {
             $store = Mage::app()->getStore();
         }
-        $currencyCode = new NostoCurrencyCode($store->getBaseCurrencyCode());
-        foreach ($quoteItems as $quoteItem) {
+
+        foreach ($quoteItems as $item) {
             /** @var Nosto_Tagging_Model_Meta_Cart_Item $model */
-            $model = Mage::getModel('nosto_tagging/meta_cart_item');
-            $model->loadData($quoteItem, $currencyCode);
-            $this->_lineItems[] = $model;
+            $this->_lineItems[] = Mage::getModel(
+                'nosto_tagging/meta_cart_item',
+                array(
+                    'productId' => (int)$this->buildProductId($item),
+                    'quantity' => (int)$item->getQty(),
+                    'name' => $this->buildProductName($item),
+                    'unitPrice' => new NostoPrice($item->getBasePriceInclTax()),
+                    'currency' => new NostoCurrencyCode($store->getBaseCurrencyCode()),
+                )
+            );
         }
+    }
+
+    /**
+     * Returns the product id for a quote item.
+     * Always try to find the "parent" product ID if the product is a child of
+     * another product type. We do this because it is the parent product that
+     * we tag on the product page, and the child does not always have it's own
+     * product page. This is important because it is the tagged info on the
+     * product page that is used to generate recommendations and email content.
+     *
+     * @param Mage_Sales_Model_Quote_Item $item the quote item model.
+     *
+     * @return int|string
+     */
+    protected function buildProductId(Mage_Sales_Model_Quote_Item $item)
+    {
+        $parentItem = $item->getOptionByCode('product_type');
+        if (!is_null($parentItem)) {
+            return $parentItem->getProductId();
+        } elseif ($item->getProductType() === Mage_Catalog_Model_Product_Type::TYPE_SIMPLE) {
+            /** @var Mage_Catalog_Model_Product_Type_Configurable $model */
+            $model = Mage::getModel('catalog/product_type_configurable');
+            $parentIds = $model->getParentIdsByChild($item->getProductId());
+            $attributes = $item->getBuyRequest()->getData('super_attribute');
+            // If the product has a configurable parent, we assume we should tag
+            // the parent. If there are many parent IDs, we are safer to tag the
+            // products own ID.
+            if (count($parentIds) === 1 && !empty($attributes)) {
+                return $parentIds[0];
+            }
+        }
+        return $item->getProductId();
+    }
+
+    /**
+     * Returns the name for a quote item.
+     * Configurable products will have their chosen options added to their name.
+     * Bundle products will have their chosen child product names added.
+     * Grouped products will have their parent product name prepended.
+     * All others will have their own name only.
+     *
+     * @param Mage_Sales_Model_Quote_Item $item the quote item model.
+     *
+     * @return string
+     */
+    protected function buildProductName(Mage_Sales_Model_Quote_Item $item)
+    {
+        switch ($item->getProductType()) {
+            case Mage_Catalog_Model_Product_Type::TYPE_SIMPLE:
+                return $this->fetchSimpleProductName($item);
+
+            case Mage_Catalog_Model_Product_Type::TYPE_CONFIGURABLE:
+                return $this->fetchConfigurableProductName($item);
+
+            case Mage_Catalog_Model_Product_Type::TYPE_BUNDLE:
+                return $this->fetchBundleProductName($item);
+
+            case Mage_Catalog_Model_Product_Type::TYPE_GROUPED:
+                return $this->fetchGroupedProductName($item);
+
+            default:
+                return $item->getName();
+        }
+    }
+
+    /**
+     * Returns the name for an quote item representing a simple product.
+     *
+     * @param Mage_Sales_Model_Quote_Item $item the quote item model.
+     *
+     * @return string
+     */
+    protected function fetchSimpleProductName(Mage_Sales_Model_Quote_Item $item)
+    {
+        $name = $item->getName();
+        $nameOptions = array();
+
+        /** @var Mage_Catalog_Model_Product_Type_Configurable $model */
+        $model = Mage::getModel('catalog/product_type_configurable');
+        $parentIds = $model->getParentIdsByChild($item->getProductId());
+        // If the product has a configurable parent, we assume we should tag
+        // the parent. If there are many parent IDs, we are safer to tag the
+        // products own name alone.
+        if (count($parentIds) === 1) {
+            $attributes = $item->getBuyRequest()->getData('super_attribute');
+            if (is_array($attributes) && count($attributes) > 0) {
+                $nameOptions = $this->getAttributeLabels($attributes);
+            }
+        }
+
+        return $this->applyProductNameOptions($name, $nameOptions);
+    }
+
+    /**
+     * Returns the name for an quote item representing a configurable product.
+     *
+     * @param Mage_Sales_Model_Quote_Item $item the quote item model.
+     *
+     * @return string
+     */
+    protected function fetchConfigurableProductName(Mage_Sales_Model_Quote_Item $item)
+    {
+        $name = $item->getName();
+        $nameOptions = array();
+
+        /* @var $helper Mage_Catalog_Helper_Product_Configuration */
+        $helper = Mage::helper('catalog/product_configuration');
+        foreach ($helper->getConfigurableOptions($item) as $opt) {
+            if (isset($opt['value']) && is_string($opt['value'])) {
+                $nameOptions[] = $opt['value'];
+            }
+        }
+
+        return $this->applyProductNameOptions($name, $nameOptions);
+    }
+
+    /**
+     * Returns the name for an quote item representing a bundle product.
+     *
+     * @param Mage_Sales_Model_Quote_Item $item the quote item model.
+     *
+     * @return string
+     */
+    protected function fetchBundleProductName(Mage_Sales_Model_Quote_Item $item)
+    {
+        $name = $item->getName();
+        $nameOptions = array();
+
+        $type = $item->getProduct()->getTypeInstance(true);
+        $opts = $type->getOrderOptions($item->getProduct());
+        if (isset($opts['bundle_options']) && is_array($opts['bundle_options'])) {
+            foreach ($opts['bundle_options'] as $opt) {
+                if (isset($opt['value']) && is_array($opt['value'])) {
+                    foreach ($opt['value'] as $val) {
+                        $qty = '';
+                        if (isset($val['qty']) && is_int($val['qty'])) {
+                            $qty .= $val['qty'] . ' x ';
+                        }
+                        if (isset($val['title']) && is_string($val['title'])) {
+                            $nameOptions[] = $qty . $val['title'];
+                        }
+                    }
+                }
+            }
+        }
+
+        return $this->applyProductNameOptions($name, $nameOptions);
+    }
+
+    /**
+     * Returns the name for an quote item representing a grouped product.
+     *
+     * @param Mage_Sales_Model_Quote_Item $item the quote item model.
+     *
+     * @return string
+     */
+    protected function fetchGroupedProductName(Mage_Sales_Model_Quote_Item $item)
+    {
+        $name = $item->getName();
+
+        $config = $item->getBuyRequest()->getData('super_product_config');
+        if (isset($config['product_id'])) {
+            /** @var Mage_Catalog_Model_Product $parent */
+            $parent = Mage::getModel('catalog/product')
+                ->load($config['product_id']);
+            $parentName = $parent->getName();
+            if (!empty($parentName)) {
+                $name = $parentName . ' - ' . $name;
+            }
+        }
+
+        return $name;
+    }
+
+    /**
+     * Returns a list of attribute labels based on given attribute option map.
+     *
+     * The map must be passed with attribute id's as keys and the option id's
+     * as values.
+     *
+     * @param array $attributes the attribute id map.
+     *
+     * @return array
+     */
+    protected function getAttributeLabels(array $attributes)
+    {
+        $labels = array();
+        if (count($attributes) > 0) {
+            /** @var Mage_Eav_Model_Entity_Attribute[] $collection */
+            $collection = Mage::getModel('eav/entity_attribute')
+                ->getCollection()
+                ->addFieldToFilter(
+                    'attribute_id',
+                    array(
+                        'in' => array_keys($attributes)
+                    )
+                );
+            foreach ($collection as $attribute) {
+                $optionId = $attributes[$attribute->getId()];
+                if (!$attribute->getData('source_model')) {
+                    $attribute->setData(
+                        'source_model',
+                        'eav/entity_attribute_source_table'
+                    );
+                }
+                try {
+                    $label = $attribute->getSource()->getOptionText($optionId);
+                    if (!empty($label)) {
+                        $labels[] = $label;
+                    }
+                } catch (Mage_Core_Exception $e) {
+                    // If the source model cannot be found, just continue;
+                    continue;
+                }
+
+            }
+        }
+        return $labels;
+    }
+
+    /**
+     * Applies given options to the name.
+     *
+     * Format:
+     *
+     * "Product Name (Green, M)"
+     *
+     * @param string $name the name.
+     * @param array  $options list of string values to apply as name option.
+     *
+     * @return string
+     */
+    protected function applyProductNameOptions($name, array $options)
+    {
+        if (!empty($options)) {
+            $name .= ' (' . implode(', ', $options) . ')';
+        }
+        return $name;
     }
 
     /**