+++ CHANGELOG.md
@@ -7,6 +7,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Multi-currency + exchange rate support
 - Account synchronisation support
 
+## [2.6.5]
+### Added
+- "external_order_ref" to order tagging and API requests in order to better
+track orders
+- "order_statuses" to order export in order to better track orders
+
 ## [2.6.4]
 ### Fixed
 - Redirect to admin store after OAuth is finished
@@ -187,7 +193,8 @@ Catalog Category" option was enabled
 ### Added
 - Initial release
 
-[unreleased]: https://github.com/nosto/nosto-magento-extension/compare/2.6.4...develop
+[unreleased]: https://github.com/nosto/nosto-magento-extension/compare/2.6.5...develop
+[2.6.5]: https://github.com/nosto/nosto-magento-extension/compare/2.6.4...2.6.5
 [2.6.4]: https://github.com/nosto/nosto-magento-extension/compare/2.6.3...2.6.4
 [2.6.3]: https://github.com/nosto/nosto-magento-extension/compare/2.6.2...2.6.3
 [2.6.2]: https://github.com/nosto/nosto-magento-extension/compare/2.6.1...2.6.2